GPII-4100: Further pared down listener teardown example.
