+++ tests/browser-fixtures/all-tests.html
@@ -14,8 +14,6 @@
             QUnit.testSuites([
                 "/dispatcher/initblock?myvar=bar",
                 "/tests/browser-fixtures/tests-hasRequiredOptions.html",
-                // TODO: Remove this once we know what the issue is.
-                "/tests/browser-fixtures/tests-listener-problems.html",
                 "/tests/browser-fixtures/tests-renderer-serverResourceAware.html",
                 "/tests/browser-fixtures/tests-renderer-standalone.html",
                 "/tests/browser-fixtures/tests-serverMessageAware.html",
