--- tests/browser-fixtures/tests-listener-problems.html
@@ -1,30 +1,0 @@
-<html>
-    <head>
-        <title>Demonstrating listener teardown problems.</title>
-
-        <!-- Fluid dependencies -->
-        <link rel="stylesheet" media="screen" href="/node_modules/infusion/tests/lib/qunit/css/qunit.css" />
-        <script type="text/javascript" src="/node_modules/infusion/dist/infusion-all.js"></script>
-
-        <!-- Required for Fluid IoC tests. -->
-        <script type="text/javascript" src="/node_modules/infusion/tests/test-core/utils/js/IoCTestUtils.js"></script>
-
-        <!-- QUnit/jqUnit and Testem dependencies -->
-        <link rel="stylesheet" media="screen" href="/node_modules/infusion/tests/lib/qunit/css/qunit.css" />
-        <script type="text/javascript" src="/node_modules/infusion/tests/lib/qunit/js/qunit.js"></script>
-        <script type="text/javascript" src="/node_modules/infusion/tests/test-core/jqUnit/js/jqUnit.js"></script>
-        <script src="/testem.js"></script>
-        <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
-    </head>
-    <body>
-        <h1 id="qunit-header">"Listener Problem" Tests</h1>
-        <h2 id="qunit-banner"></h2>
-        <div id="qunit-testrunner-toolbar"></div>
-        <h2 id="qunit-userAgent"></h2>
-        <ol id="qunit-tests"></ol>
-
-        <div class="container"></div>
-        <!-- The tests -->
-        <script src="./js/tests-listener-problems.js"></script>
-    </body>
-</html>
